Support for hosting addon in a subpath
